Fixed sending UDP packets
By mistake my previous commit included the original RtpSocket class which was streaming via Multicast (didn't work for me).
This commit includes the modified versions for Rosetta which uses Unicast instead.

I also moved the H264Packetizer code that streams the data into a separate thread. Previously we were just calling run() instead of starting the thread. I also added some checks to avoid changing the buffer while it is being sent by the other thread, but this was not happening anyway.

I also added a boolean "splitNALs" that can be modified using the debugger to disable the NAL splitting.
I discovered that disabling this boolean after the client started to receive the stream, improved the quality significantly, which means that we are probably doing something wrong in the NAL splitting code.
Without NAL splitting, gstreamer was not able to open the stream.